Retire `v1beta1` version of `sdk`
This version is not used anymore after we have merged
this PR with significant improvements https://github.com/volur-ai/python-volur-ai-sdk/pull/41
