+++ src/volur/sdk/__init__.py
@@ -1,21 +1,9 @@
 """A package where all SDK-related code lives"""
 
-from volur.sdk.v1alpha1 import (
-    CharacteristicColumn,
-    Column,
-    MaterialsCSVFileSource,
-    MaterialsSource,
-    QuantityColumn,
-    Value,
+from .v1alpha2 import (
     VolurClient,
 )
 
 __all__ = [
-    "CharacteristicColumn",
-    "Column",
-    "MaterialsSource",
-    "MaterialsCSVFileSource",
-    "QuantityColumn",
-    "Value",
     "VolurClient",
 ]
