+++ src/volur/sdk/v1alpha2/__init__.py
@@ -1,4 +1,4 @@
-from volur.sdk.v1alpha2.client import VolurClient
+from .client import VolurClient
 
 __all__ = [
     "VolurClient",
