+++ src/volur/sdk/v1alpha2/sources/__init__.py
@@ -1,1 +1,5 @@
+from .csv import MaterialsSource
 
+__all__ = [
+    "MaterialsSource",
+]
