--- tests/test_sources/csv/v1alpha2/test_characteristic_column.py
@@ -1,1 +1,0 @@
-
