+++ tests/test_sources/csv/v1alpha2/test_column.py
@@ -1,5 +1,5 @@
 import pytest
-from volur.sdk.v1alpha2.sources.csv.base import Column
+from volur.sdk.v1alpha2.sources.csv import Column
 
 test_ids = [
     "pass-with-a-column-name",
